Fixed test paths and names
