+++ src/test/java/com/uid2/core/util/OperatorInfoTest.java
