+++ src/test/java/com/uid2/core/vertx/ServiceLinkMetadataPathTest.java
@@ -36,14 +36,13 @@ import java.net.URL;
 import java.util.HashSet;
 
 import static com.uid2.shared.Utils.readToEndAsString;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(VertxExtension.class)
-public class TestClientSideKeypairMetadataPath {
-
+public class ServiceLinkMetadataPathTest {
     @Mock
     private ICloudStorage cloudStorage;
     @Mock
@@ -95,16 +94,16 @@ public class TestClientSideKeypairMetadataPath {
 
     @Test
     void publicOperatorGetsGlobalKeypairs(Vertx vertx, VertxTestContext testContext) throws CloudStorageException, IOException {
-        String metadata = "/com.uid2.core/testGlobalMetadata/client_side_keypairs/metadata.json";
+        String metadata = "/com.uid2.core/testGlobalMetadata/service_links/metadata.json";
         String metadataContent = openFile(metadata);
-        String location = ((JsonObject) Json.decodeValue(metadataContent)).getJsonObject("client_side_keypairs").getString("location");
+        String location = ((JsonObject) Json.decodeValue(metadataContent)).getJsonObject("service_links").getString("location");
         when(cloudStorage.download(eq(metadata))).thenReturn(new StringBufferInputStream(metadataContent));
         when(cloudStorage.preSignUrl(any())).thenAnswer(i -> new URL(i.getArgument(0)));
         fakeAuth(OperatorType.PUBLIC, 99);
-        get(vertx, "/client_side_keypairs/refresh", "", testContext.succeeding(response -> testContext.verify(()-> {
+        get(vertx, "/service_links/refresh", "", testContext.succeeding(response -> testContext.verify(() -> {
             assertEquals(200, response.statusCode());
             JsonObject json = response.bodyAsJsonObject();
-            String resultLocation = json.getJsonObject("client_side_keypairs").getString("location");
+            String resultLocation = json.getJsonObject("service_links").getString("location");
             assertEquals(resultLocation, location);
             testContext.completeNow();
         })));
@@ -112,21 +111,20 @@ public class TestClientSideKeypairMetadataPath {
 
     @Test
     void privateOperatorGetsKeypairsError(Vertx vertx, VertxTestContext testContext) throws CloudStorageException, IOException {
-        String metadata = "/com.uid2.core/testGlobalMetadata/client_side_keypairs/metadata.json";
+        String metadata = "/com.uid2.core/testGlobalMetadata/service_links/metadata.json";
         String metadataContent = openFile(metadata);
-        String location = ((JsonObject) Json.decodeValue(metadataContent)).getJsonObject("client_side_keypairs").getString("location");
+        String location = ((JsonObject) Json.decodeValue(metadataContent)).getJsonObject("service_links").getString("location");
         when(cloudStorage.download(eq(metadata))).thenReturn(new StringBufferInputStream(metadataContent));
         when(cloudStorage.preSignUrl(any())).thenAnswer(i -> new URL(i.getArgument(0)));
         fakeAuth(OperatorType.PRIVATE, 99);
-        get(vertx, "/client_side_keypairs/refresh", "", testContext.succeeding(response -> testContext.verify(() -> {
+        get(vertx, "/service_links/refresh", "", testContext.succeeding(response -> testContext.verify(() -> {
             assertEquals(403, response.statusCode());
-            assertEquals("endpoint /client_side_keypairs/refresh is for public operators only", response.bodyAsJsonObject().getString("message"));
+            assertEquals("endpoint /service_links/refresh is for public operators only", response.bodyAsJsonObject().getString("message"));
             testContext.completeNow();
         })));
     }
 
-    String openFile(String filePath) throws IOException
-    {
-        return readToEndAsString(TestSiteSpecificMetadataPath.class.getResourceAsStream(filePath));
+    String openFile(String filePath) throws IOException {
+        return readToEndAsString(SiteSpecificMetadataPathTest.class.getResourceAsStream(filePath));
     }
 }
